Fix the fix (r11718) for bug #362.
We were still waiting the entire duration of the timeout before we figured out that a connect() was successful.  Re-introduce adding the peer_send_event so that we detect immediately when a connect() completes.

Also make sure to delete the timeout event in complete_connect().

Fixed a struct timeval initialization warning reported by Jeff.

Remove an erroneous opal_output().



git-svn-id: http://svn.open-mpi.org/svn/ompi/trunk@11724 63e3feb5-37d5-0310-a306-e8a459e722fe