Fix for bug #369.
LoadLeveler only sets LOADL_PROCESSOR_LIST when there are 128 or less tasks allocated to a job. The POE RAS relied on this variable so I created a new RAS which uses the LoadLeveler API instead of relying on the environment variable. This still needs some testing, so for now we use the POE RAS whenever LOADL_PROCESSOR_LIST, otherwise we fall back on this component.

Unfortunately, this will require an autogen...


git-svn-id: http://svn.open-mpi.org/svn/ompi/trunk@11732 63e3feb5-37d5-0310-a306-e8a459e722fe