Clean up an error in the mapper that caused "-hosts" to bomb.
Update the mapper so it correctly points to the next node to be used if we are mapping by slots. As it was, if we had an app_context that used only one slot on a node, the next app_context would start on the next node - leaving a blank slot in-between.


git-svn-id: http://svn.open-mpi.org/svn/ompi/trunk@12193 63e3feb5-37d5-0310-a306-e8a459e722fe