Rewriting much of the errorcode and errorclass code, since  - we have to be able to attach a string to an error class, not just to an  error code  - according to MPI-2 the attribute MPI_LASTUSEDCODE has to be updated   everytime you add a new code or a new class. Thus, you have to have single   list for both.
Thus, we got rid of the error_class structure. In the error-code structure, we
can distinguish whether we are dealing with an error code or an error class by
looking at the err->code element of the structure. In case its value is
MPI_UNDEFINED, the according entry is a class, else it is an error code. All
predefined error codes have the code and the class field set to the same
value.

The test MPI_Add_error_class1 passes now.

Fixes #418



git-svn-id: http://svn.open-mpi.org/svn/ompi/trunk@12764 63e3feb5-37d5-0310-a306-e8a459e722fe