Add a check for the unexpected handler. If enabled, allow the zero-copy protocol over the MX BTL. Now, we have only one matching, the one in Open MPI.
The problem is that when the unexpected handler is triggered, not all the
message is on the host memory. In the best case we get one MX fragment (internal
MX fragment), in the worst we get NULL. The only way to fit this with the
design of the PML is to force the eager protocol at the MX internal fragment
size, and to limit the send/receive protocol at the same size. Tests show
the outcome is not far from optimal (if the pipeline depth is increased
a little bit).

Set MX_PIPELINE_LOG in order to allow MX to use internal fragments of 4K.


git-svn-id: http://svn.open-mpi.org/svn/ompi/trunk@12930 63e3feb5-37d5-0310-a306-e8a459e722fe