Refs #853
Add new function opal_get_num_processors() that will return the number
of processors on the local host.  Does the Right thing in POSIX
environments (to include a special case for OS X), and will shortly do
the Right Thing for Windows (this commit includes a change to
configure, so I wanted to get that in before the US workday -- the
Windows code can some shortly because it won't involve configury
changes).


git-svn-id: http://svn.open-mpi.org/svn/ompi/trunk@13506 63e3feb5-37d5-0310-a306-e8a459e722fe