Submitted by Gleb, reviewed by Rich:
Queue_empty is determined by the reader, and is it's local view.
However, the writer may continue writing to this queue.  The decision
to go on to the next cb_fifo is done in an atomic region, checking the
writer's view.  The writer also "changes it's view" in an atomic
region protected by the same lock.


git-svn-id: http://svn.open-mpi.org/svn/ompi/trunk@13968 63e3feb5-37d5-0310-a306-e8a459e722fe