Add some basic timing hooks so I can extract a few more detailed performance numbers for tuning.
Switch the bookmark_recv to be non-blocking. If this is blocking then for
process counts >= 32 slight process delays were causing cascading performance
delays in the protocol. This lead to checkpoints either taking about 3 sec or
45 sec (or more) for 64 procs due to the cascading delays. With the nonblocking
receive version this is no longer the case we get the speedup we expect for this
part of the protocol.

More tuning to come.



git-svn-id: http://svn.open-mpi.org/svn/ompi/trunk@16137 63e3feb5-37d5-0310-a306-e8a459e722fe