'''TEMPORARY''' fix for #1211 to allow normal builds on the trunk: ensure that including valgrind.h is within the WANT_MEMCHECKER macro.
But this does not seem right; memchecker is a framework and valgrind
is the component -- why are we including a component-specific header
file in a top-level OMPI .h file?  It seems like there should be a
memchecker .h file that generically hides these component-specific
details (i.e., what if we add some other memchecker components
someday?).


git-svn-id: http://svn.open-mpi.org/svn/ompi/trunk@17432 63e3feb5-37d5-0310-a306-e8a459e722fe