From Brian B.:
This commit lowers the priority of the darwin backtrace component
below that of the ''execinfo'' and ''stackprint'' components, which
will cause OS X Leopard to use the ''execinfo'' component.  execinfo
utilizes a public API for printing the stacktrace.  The ''darwin''
component uses some evil hacks and a not-so supported package from
Apple to print the stack trace.  


git-svn-id: http://svn.open-mpi.org/svn/ompi/trunk@17840 63e3feb5-37d5-0310-a306-e8a459e722fe