Sorry, previous commit was in the wrong directory. This is the real fix (have to undo 1822).
The verification of recvcount==0 and rank = root was braking
inter-communicator scatter, since the root (root==MPI_ROOT) might very well
have recvcount=0. The same fix has been applied to gather.c just the other way
round. 
 
Fixes the bug reported on the mainling list by Martin Audet. If there is a
1.2.7 this fix might be worthwhile porting it over.

Please note, that while the test works now for basic and for inter, we get a
0byte malloc warning from the inter module, which we still have to fix in a
separate patch.




git-svn-id: http://svn.open-mpi.org/svn/ompi/trunk@18123 63e3feb5-37d5-0310-a306-e8a459e722fe